Update .travis.yml
### Travis messages
#### --dev flag: 
`You are using the deprecated option "dev". Dev packages are installed by default now.` I removed the flag

#### phpunit/phpunit-mock-objects  package:
`Package phpunit/phpunit-mock-objects is *abandoned*, you should avoid using it. No replacement was suggested.`

### Other changes
#### dpkg:
To be able to use --allow-unauthenticated flag to bypass mongodb Error.

#### -y:
Use the -y parameter with apt-get to assume yes to all queries by the apt tools. Postfix was not installed because -y was not available.

####: -q
Quite.